changes to ValidationConfig: - move from config.py to validation_config.py - add to __init__.py for gamechangerml/configs - update references to this class
