+++ gamechangerml/configs/__init__.py
@@ -7,3 +7,4 @@ from .doc_compare_similarity_config import DocCompareSimilarityConfig
 from .qexp_config import QexpConfig
 from .training_config import TrainingConfig
 from .d2v_config import D2VConfig
+from .validation_config import ValidationConfig
